+++ react/FetchMore.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { path } from 'ramda'
 import classNames from 'classnames'
+import PropTypes from 'prop-types'
 import FetchMoreButton from './components/loaders/FetchMoreButton'
 import LoadingSpinner from './components/loaders/LoadingSpinner'
 import { PAGINATION_TYPE } from './constants/paginationType'
@@ -10,7 +11,7 @@ import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
 
 import styles from './searchResult.css'
 
-const FetchMore = () => {
+const FetchMore = ({ htmlElementForButton = 'button '}) => {
   const { pagination, searchQuery, maxItemsPerPage, page } = useSearchPage()
   const products = path(['data', 'productSearch', 'products'], searchQuery)
   const recordsFiltered = path(
@@ -51,12 +52,18 @@ const FetchMore = () => {
           onFetchMore={handleFetchMoreNext}
           loading={loading}
           showProductsCount={false}
+          htmlElementForButton={htmlElementForButton}
         />
       </div>
     )
   }
 
   return <LoadingSpinner loading={loading} />
+}
+
+FetchMore.propTypes = {
+  /* html element to render for fetch more button */
+  htmlElementForButton: PropTypes.string,
 }
 
 export default FetchMore