+++ react/FetchPrevious.js
@@ -1,13 +1,14 @@
 import React from 'react'
 import { path } from 'ramda'
 import classNames from 'classnames'
+import PropTypes from 'prop-types'
 import FetchPreviousButton from './components/loaders/FetchPreviousButton'
 import { useFetchMore } from './hooks/useFetchMore'
 import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
 
 import styles from './searchResult.css'
 
-const FetchPrevious = () => {
+const FetchPrevious = ({ htmlElementForButton = 'button' }) => {
   const { searchQuery, maxItemsPerPage, page } = useSearchPage()
   const products = path(['data', 'productSearch', 'products'], searchQuery)
   const recordsFiltered = path(
@@ -45,9 +46,15 @@ const FetchPrevious = () => {
         recordsFiltered={recordsFiltered}
         onFetchPrevious={handleFetchMorePrevious}
         loading={loading}
+        htmlElementForButton={htmlElementForButton}
       />
     </div>
   )
+}
+
+FetchPrevious.propTypes = {
+  /* html element to render for fetch previous button */
+  htmlElementForButton: PropTypes.string,
 }
 
 export default FetchPrevious